Function gpsync::main

pub(crate) fn main() -> Result<(), &'static str>

sync WDInfo tables from PostgreSQL into Greenplum in docker

Greenplum in docker start (maybe merge then into docker and rust code later):

ssh -i .ssh/ubuntu_sinnud.pem sinnud@192.168.1.213

cd /var/local/docker/greenplum-oss-docker/gpdb

sudo ./run_ossdocker.sh

startGPDB.sh

su - gpadmin

psql

create user sinnud WITH PASSWORD 'password';

create database mydb;

GRANT ALL PRIVILEGES ON DATABASE mydb TO sinnud;

ALTER USER sinnud WITH ENCRYPTED PASSWORD 'password';

\q

psql -U sinnud -h localhost -d mydb

create schema wdinfo;

\q

After investigating docker, we can reduce the above command as:

ssh -i .ssh/ubuntu_sinnud.pem sinnud@192.168.1.213

cd /var/local/docker/greenplum-oss-docker/gpdb

make sure we have scripts folder with two files gpinit.bash and gpinit2.bash (see contents below)

sudo ./run_ossdocker.sh &

sudo docker exec -it gpdb5oss startGPDB.sh

sudo docker exec -it gpdb5oss su - gpadmin /code/scripts/gpinit.bash

sudo docker exec -it gpdb5oss su - gpadmin /code/scripts/gpinit2.bash

content of gpinit.bash and gpinit2.bash

script file: gpinit.bash
psql -c "create user sinnud WITH PASSWORD 'password';"
psql -c "create database mydb;"
psql -c "GRANT ALL PRIVILEGES ON  DATABASE mydb TO sinnud;"
psql -c "ALTER USER sinnud WITH ENCRYPTED PASSWORD 'password';"

script file: gpinit2.bash
PGPASSWORD=password psql -U sinnud -h localhost -d mydb -c "create schema wdinfo;"